libraries/Core/microsoft-agents-authentication/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "microsoft-agents-authentication"
version = "0.0.0a1"
description = "An authentication library for Microsoft Agents"
authors = [{name = "Microsoft Corporation"}]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyjwt>=2.10.1"
]

[project.urls]
"Homepage" = "https://github.com/microsoft/microsoft-agents-protocol"
